# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "HandyConnect API documentation",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:10000",
    "basePath": "/",
    "paths": {
        "/": {
            "get": {
                "description": "get HomePage",
                "tags": [
                    "Home"
                ],
                "summary": "Get HomePage",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/availability": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "get service's availability in a city",
                "parameters": [
                    {
                        "type": "string",
                        "description": "city ID",
                        "name": "city_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "service ID",
                        "name": "service_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/structTypes.Date"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/booking": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "create new service booking",
                "parameters": [
                    {
                        "description": "booking data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.BookingSchema"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/cities": {
            "get": {
                "description": "get all cities in which services are available",
                "summary": "Get all cities",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/structTypes.City"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/customer": {
            "get": {
                "description": "Get Customer by ID",
                "summary": "Get Customer data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "customer id",
                        "name": "customer_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/structTypes.Cust"
                        }
                    }
                }
            }
        },
        "/customerLogin": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "verify email and password of customer",
                "parameters": [
                    {
                        "description": "Login data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/structTypes.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/customerRating": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "rating provided by customer for a booking",
                "parameters": [
                    {
                        "description": "rating data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/structTypes.RatingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/customerSignUp": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "create new customer",
                "parameters": [
                    {
                        "description": "customer data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.CustomerProfileSchema"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/vendor": {
            "get": {
                "description": "Get Vendor by ID",
                "summary": "Get Vendor data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "vendor_id",
                        "name": "vendor_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/structTypes.Vendor"
                        }
                    }
                }
            }
        },
        "/vendorLogin": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "verify email and password of vendor",
                "parameters": [
                    {
                        "description": "Login data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/structTypes.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/vendorRating": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "rating provided by vendor for a customer",
                "parameters": [
                    {
                        "description": "rating data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/structTypes.RatingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/vendorSignUp": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "create new vendor",
                "parameters": [
                    {
                        "description": "vendor data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.VendorProfileSchema"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "schema.BookingSchema": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "booking_status": {
                    "type": "string"
                },
                "city_id": {
                    "type": "integer"
                },
                "customer_id": {
                    "type": "integer"
                },
                "customer_rating": {
                    "type": "integer"
                },
                "day": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "month": {
                    "type": "integer"
                },
                "service_id": {
                    "type": "integer"
                },
                "vendor_id": {
                    "type": "integer"
                },
                "vendor_rating": {
                    "type": "integer"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "schema.CustomerProfileSchema": {
            "type": "object",
            "properties": {
                "city_id": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "rating": {
                    "type": "string"
                },
                "rating_count": {
                    "type": "string"
                }
            }
        },
        "schema.VendorProfileSchema": {
            "type": "object",
            "properties": {
                "city_id": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "rating": {
                    "type": "string"
                },
                "rating_count": {
                    "type": "string"
                },
                "service1_id": {
                    "type": "integer"
                },
                "service2_id": {
                    "type": "integer"
                },
                "service3_id": {
                    "type": "integer"
                }
            }
        },
        "structTypes.City": {
            "type": "object",
            "properties": {
                "city_id": {
                    "type": "integer"
                },
                "city_name": {
                    "type": "string"
                }
            }
        },
        "structTypes.Cust": {
            "type": "object",
            "properties": {
                "city_id": {
                    "type": "integer"
                },
                "city_name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string"
                },
                "phone": {
                    "type": "integer"
                },
                "rating": {
                    "type": "number"
                },
                "rating_count": {
                    "type": "integer"
                }
            }
        },
        "structTypes.Date": {
            "type": "object",
            "properties": {
                "day": {
                    "type": "integer"
                },
                "month": {
                    "type": "integer"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "structTypes.LoginRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "structTypes.RatingRequest": {
            "type": "object",
            "properties": {
                "booking_id": {
                    "type": "integer"
                },
                "rating": {
                    "type": "integer"
                }
            }
        },
        "structTypes.Vendor": {
            "type": "object",
            "properties": {
                "city_name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string"
                },
                "phone": {
                    "type": "integer"
                },
                "rating": {
                    "description": "Service2   string `json: \"service2\"`\nService3   string `json: \"service3\"`",
                    "type": "number"
                },
                "rating_count": {
                    "type": "integer"
                },
                "service_name": {
                    "type": "string"
                }
            }
        }
    }
}